+++ src/Definition/Plugin/AbstractDefinitionPlugin.php
@@ -8,14 +8,14 @@
  *  file that was distributed with this source code.
  ******************************************************************************/
 
-namespace Ynlo\GraphQLBundle\Definition\Extension;
+namespace Ynlo\GraphQLBundle\Definition\Plugin;
 
 use Doctrine\Common\Util\Inflector;
 use Symfony\Component\Config\Definition\Builder\ArrayNodeDefinition;
 use Ynlo\GraphQLBundle\Definition\DefinitionInterface;
 use Ynlo\GraphQLBundle\Definition\Registry\Endpoint;
 
-abstract class AbstractDefinitionExtension implements DefinitionExtensionInterface
+abstract class AbstractDefinitionPlugin implements DefinitionPluginInterface
 {
     /**
      * {@inheritDoc}
@@ -24,7 +24,7 @@ abstract class AbstractDefinitionExtension implements DefinitionExtensionInterfa
     {
         $name = \get_class($this);
         preg_match('/(\w+)$/', $name, $matches);
-        $name = preg_replace('/Extension$/', null, $matches[1]);
+        $name = preg_replace('/Plugin/', null, $matches[1]);
         $name = preg_replace('/Definition$/', null, $name);
 
         return Inflector::tableize($name);