+++ src/Definition/Plugin/ReorderDefinitionsPlugin.php
@@ -8,7 +8,7 @@
  *  file that was distributed with this source code.
  ******************************************************************************/
 
-namespace Ynlo\GraphQLBundle\Definition\Extension;
+namespace Ynlo\GraphQLBundle\Definition\Plugin;
 
 use Ynlo\GraphQLBundle\Definition\MutationDefinition;
 use Ynlo\GraphQLBundle\Definition\QueryDefinition;
@@ -17,7 +17,7 @@ use Ynlo\GraphQLBundle\Definition\Registry\Endpoint;
 /**
  * Reorder queries & mutations by name based on related node
  */
-class ReorderDefinitionsExtension extends AbstractDefinitionExtension
+class ReorderDefinitionsPlugin extends AbstractDefinitionPlugin
 {
     /**
      * {@inheritDoc}